Move all audio manager files into TizenRT/framework
